add token auth

+++ service/task.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from exception import TaskNotFoundException
 from repository import TaskRepository, TaskCache
 from schema import TaskSchema, TaskCreateSchema
 
@@ -19,8 +20,17 @@ class TaskService:
         return tasks
     
 
-    def get_task(self, task_id: int) -> TaskSchema:
+    def get_task(self, task_id: int, user_id: int) -> TaskSchema:
+        task = self.task_repository.get_user_task(
+            task_id=task_id,
+            user_id=user_id
+        )
+
+        if not task:
+            raise TaskNotFoundException
+
         task = self.task_repository.get_task(task_id)
+
         return task
     
 
@@ -29,12 +39,28 @@ class TaskService:
 
         return self.get_task(task_id)
     
-    def delete_task(self, task_id):
+    def delete_task(self, task_id: int, user_id: int):
+        task = self.task_repository.get_user_task(
+            task_id=task_id,
+            user_id=user_id
+        )
+
+        if not task:
+            raise TaskNotFoundException
+
         self.task_repository.delete_task(task_id)
 
 
-    def update_task(self, task_id: int, name: str) -> TaskSchema:
-        task = self.task_repository.update_name(task_id, name)
+    def update_task(self, task_id: int, task_new_name: str, user_id: int) -> TaskSchema:
+        task = self.task_repository.get_user_task(
+            task_id=task_id,
+            user_id=user_id
+        )
+
+        if not task:
+            raise TaskNotFoundException
+        
+        task = self.task_repository.update_name(task_id, name=task_new_name)
         
         return task
